Rename module
Now that `Channels` has been moved out, `channel` fits much better.
